+++ src/utils/__tests__/isNumbery.test.js
@@ -1,0 +1,11 @@
+import isNumbery from "../isNumbery"
+
+it("isNumbery", () => {
+  expect(isNumbery("1")).toBeTruthy()
+  expect(isNumbery("21 ")).toBeTruthy()
+  expect(isNumbery(" 212")).toBeTruthy()
+  expect(isNumbery("232 ")).toBeTruthy()
+  expect(isNumbery("")).toBeFalsy()
+  expect(isNumbery(" ")).toBeFalsy()
+  expect(isNumbery("a")).toBeFalsy()
+})