+++ src/utils/__tests__/nextNonCommentNode.test.js
@@ -1,0 +1,34 @@
+import nextNonCommentNode from "../nextNonCommentNode"
+import postcss from "postcss"
+
+it("nextNonCommentNode", () => {
+  const caseA = "a {} /* x */ b {}"
+  postcss().process(caseA).then(result => {
+    let aNode
+    let bNode
+    result.root.walkRules(rule => {
+      if (rule.selector === "a") {
+        aNode = rule
+      }
+      if (rule.selector === "b") {
+        bNode = rule
+      }
+    })
+    expect(nextNonCommentNode(aNode.next())).toBe(bNode)
+    expect(nextNonCommentNode(bNode.next())).toBe(null)
+  })
+
+  const caseB = "a { /* x */ color: pink; /* y */ }"
+  postcss().process(caseB).then(result => {
+    let aNode
+    let colorNode
+    result.root.walkRules(rule => {
+      aNode = rule
+    })
+    result.root.walkDecls(decl => {
+      colorNode = decl
+    })
+    expect(nextNonCommentNode(aNode.first)).toBe(colorNode)
+    expect(nextNonCommentNode(colorNode.next())).toBe(null)
+  })
+})